Make Test ScalaJS friendly
Doubles print differently on ScalaJS, use Ints instead.

+++ tests/run/i15725.scala
@@ -1,8 +1,8 @@
-class ToString(d:Double) extends AnyVal :
+class ToString(d: Int) extends AnyVal :
   inline override def toString():String = s"$d meters"
 
 @main def Test =
-  val ts = ToString(2.0)
+  val ts = ToString(2)
   val a: Any = ts
   println(a)
